Convert the retrieval of the active data files to streaming
Instead of waiting to get the collection in bulk of all
the active data files of a table, provide an closeable iterator
so that the split manager can start providing splits
right away.

The provided iterator is closeable in order to ensure that
the underlying checkpoint parquet page source gets closed
even when not all the entries from the checkpoint are consumed.