Strip from the dynamic filter predicate the partition columns domains
Avoid using filters which relate to the partition columns
in the predicate used in the page source because they
are irrelevant at this stage.

Moreover, if the orc or parquet data files of the table do not
have statistics for the partition columns, applying filter
predicates which relate to the partitioning columns
will produce 0 results in the filtering when the dynamic filter
is employed.