Use the actual partition name from the storage in syncing partition metadata
The canonical partition name built from the column name and
values of the partition may differ from the actual storage
location of the partition.
This can lead to inconsistencies when syncing partition
metadata, like identifying a partition to add/remove when
there is actually nothing to be done.
Rely on the partition name from the storage location for
identifying the partitions to sync via
the `sync_partition_metadata` procedure call.